--- a/libs/designer/src/lib/core/graphlayout/index.ts
+++ b/libs/designer/src/lib/core/graphlayout/index.ts
@@ -1,1 +1,1 @@
-export { useLayout } from './elklayout';
+export { useLayout, LayoutProvider } from './elklayout';