## Main Changes

Refactored our `useLayout` hook as a context provider.
I noticed some of our simple layouting changes were taking a while to update, and I found that our `useLayout` hook was recalculating individually for each instance of the hook (we had 4 separate instances).
This all happened at the same time so in addition to happening 4x instead of 1x, it seemed to bottleneck and lead to slower execution on each instance.

I've taken some perf timings for before and after my changes on just our `All Scope Nodes` test flow.

| [5 test averages] | Before | After | % Change |
|-------------------|--------|-------|----------|
| Initial load      | 1700ms | 300ms | -82%     |
| Collapse all      | 80ms   | 15ms  | -82%     |
| Expand all        | 625ms  | 50ms  | -92%     |